+++ uconnrcmpy/tests/test_experiments.py
@@ -22,6 +22,7 @@ def test_create_reactive_experiment(files):
     assert np.isclose(exp.ignition_delay, 65.81)
     assert np.isclose(exp.T_EOC, 765.374)
     assert np.isclose(exp.first_stage, 63.80)
+    assert np.isclose(exp.pressure_trace.p_EOC, 30.111)
 
 
 def test_create_nonreactive_experiment(files):
@@ -30,3 +31,4 @@ def test_create_nonreactive_experiment(files):
     assert np.isclose(exp.ignition_delay, 0.0)
     assert np.isclose(exp.T_EOC, 0.0)
     assert np.isclose(exp.first_stage, 0.0)
+    assert np.isclose(exp.pressure_trace.p_EOC, 30.184)