+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Butterworth filter is now first order
 - Butterworth filter cutoff frequency is no longer corrected
 - End point for linear fit of filter residuals is chosen automatically
+- Derivative is smoothed with moving average
 
 ### Removed